# target/web/public/main/swagger.json
{"paths":{"/api/v1/cars":{"get":{"operationId":"getAllCars","tags":["routes"]}},"/api/v1/cars/{id}":{"get":{"operationId":"getCar","tags":["routes"],"parameters":[{"in":"path","name":"id","type":"integer","format":"int64","required":true}]}},"/api/v1/cars/create":{"post":{"operationId":"createCar","tags":["routes"]}},"/api/v1/cars/update/{id}":{"put":{"operationId":"updateCar","tags":["routes"],"parameters":[{"in":"path","name":"id","type":"integer","format":"int64","required":true}]}},"/api/v1/cars/delete/{id}":{"delete":{"operationId":"deleteCar","tags":["routes"],"parameters":[{"in":"path","name":"id","type":"integer","format":"int64","required":true}]}},"/api/v1/drivers":{"get":{"operationId":"getAllDrivers","tags":["routes"]}},"/api/v1/drivers/search":{"get":{"operationId":"searchDrivers","tags":["routes"],"parameters":[{"in":"query","name":"username","type":"string","required":false,"default":"null"},{"in":"query","name":"onlineStatus","type":"string","required":false,"default":"null"},{"in":"query","name":"licencePlate","type":"string","required":false,"default":"null"},{"in":"query","name":"rating","type":"string","required":false,"default":"null"}]}},"/api/v1/drivers/{id}":{"get":{"operationId":"getDriver","tags":["routes"],"parameters":[{"in":"path","name":"id","type":"integer","format":"int64","required":true}]}},"/api/v1/drivers/create":{"post":{"operationId":"createDriver","tags":["routes"]}},"/api/v1/drivers/update/{id}":{"put":{"operationId":"updateDriver","tags":["routes"],"parameters":[{"in":"path","name":"id","type":"integer","format":"int64","required":true}]}},"/api/v1/drivers/delete/{id}":{"get":{"operationId":"deleteDriver","tags":["routes"],"parameters":[{"in":"path","name":"id","type":"integer","format":"int64","required":true}]}},"/api/v1/drivers/select-car":{"post":{"operationId":"selectCar","tags":["routes"]}},"/api/v1/drivers/deselect-car":{"post":{"operationId":"deselectCar","tags":["routes"]}},"/api/v1/authorization/generate-key":{"post":{"operationId":"generateKey","tags":["routes"]}}},"produces":["application/json"],"definitions":{},"swagger":"2.0","consumes":["application/json"],"info":{"description":"Backend REST API service implementation in Java to manage Cars and Drivers","title":"Cars and Drivers API","version":"1.0-SNAPSHOT"},"tags":[{"name":"routes"}]}